requirements.txt
cycler==0.10.0
Django==2.2
kiwisolver==1.1.0
matplotlib==3.2.0
numpy==1.18.1
Pillow==6.2.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
tqdm==4.19.9
